Connect to processes using a pipe, not a pty
Commit c1ee92e changed the yt-dl process to communicate with a pty.
However, that caused yt-dl to output escape codes for coloring it's
output, since the process now thought it was talking to a terminal.
This was fixed in commit 6de5279, which explicitly request that yt-dl
not output the escape codes. I'll leave the extra arguments in place
for now, since it doesn't hurt to be explicit.

For more info see:
https://nullprogram.com/blog/2018/01/17/
and
https://github.com/skeeto/elfeed/commit/945765a57d2f27996b6a43bc62e803dc167d1547.